pyproject.toml
[project]
name = "imagetoascii"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "nifle3", email = "kuper116rus.2005@gmail.com"},
]
dependencies = [
    "Pillow>=10.4.0",
    "numpy>=2.1.0",
    "requests>=2.32.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
test = [
]
